flambe v0.0.1 Flambe.Accounts

The Accounts context.

Functions

add_user_integration(user, attrs)

authenticate_by_email_password(email, password)

change_attention(attention)

Returns an %Ecto.Changeset{} for tracking mantra changes.

Examples

iex> change_mantra(mantra)
%Ecto.Changeset{source: %Mantra{}}

change_category(category)

Returns an %Ecto.Changeset{} for tracking category changes.

Examples

iex> change_category(category)
%Ecto.Changeset{source: %Category{}}

change_credential(credential)

Returns an %Ecto.Changeset{} for tracking credential changes.

Examples

iex> change_credential(credential)
%Ecto.Changeset{source: %Credential{}}

change_mantra(mantra)

Returns an %Ecto.Changeset{} for tracking mantra changes.

Examples

iex> change_mantra(mantra)
%Ecto.Changeset{source: %Mantra{}}

change_search_term(search_term)

Returns an %Ecto.Changeset{} for tracking search_term changes.

Examples

iex> change_search_term(search_term)
%Ecto.Changeset{source: %SearchTerm{}}

change_tabs(tabs)

Returns an %Ecto.Changeset{} for tracking tabs changes.

Examples

iex> change_tabs(tabs)
%Ecto.Changeset{source: %Tabs{}}

change_todo(todo)

Returns an %Ecto.Changeset{} for tracking todo changes.

Examples

iex> change_todo(todo)
%Ecto.Changeset{source: %Todo{}}

change_user(user)

Returns an %Ecto.Changeset{} for tracking user changes.

Examples

iex> change_user(user)
%Ecto.Changeset{source: %User{}}

create_attention(user_id, attrs \\ %{})

Creates an attenion (shift).

⚠️ example is not correct

Examples

iex> create_mantra(%{field: value})
{:ok, %Mantra{}}

iex> create_mantra(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_category(activity_ids, user_id, attrs \\ %{})

Creates a category.

Examples

iex> create_category(%{field: value})
{:ok, %Category{}}

iex> create_category(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_credential(attrs \\ %{})

Creates a credential.

Examples

iex> create_credential(%{field: value})
{:ok, %Credential{}}

iex> create_credential(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_mantra(user_id, attrs \\ %{})

Creates a mantra.

⚠️ example is not correct

Examples

iex> create_mantra(%{field: value})
{:ok, %Mantra{}}

iex> create_mantra(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_search_term(user_id, attrs \\ %{})

Creates a search_term.

Examples

iex> create_search_term(%{field: value})
{:ok, %SearchTerm{}}

iex> create_search_term(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_tabs(user_id, attrs \\ %{})

Creates a tabs.

Examples

iex> create_tabs(%{field: value})
{:ok, %Tabs{}}

iex> create_tabs(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_todo(user_id, attrs \\ %{})

Creates a todo.

Examples

iex> create_todo(%{field: value})
{:ok, %Todo{}}

iex> create_todo(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_user(attrs \\ %{})

Creates a user.

Examples

iex> create_user(%{field: value})
{:ok, %User{}}

iex> create_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_user_access_token(user)

delete_attention(attention)

Deletes a Mantra.

Examples

iex> delete_mantra(mantra)
{:ok, %Mantra{}}

iex> delete_mantra(mantra)
{:error, %Ecto.Changeset{}}

delete_category(category)

Deletes a Category.

Examples

iex> delete_category(category)
{:ok, %Category{}}

iex> delete_category(category)
{:error, %Ecto.Changeset{}}

delete_credential(credential)

Deletes a Credential.

Examples

iex> delete_credential(credential)
{:ok, %Credential{}}

iex> delete_credential(credential)
{:error, %Ecto.Changeset{}}

delete_mantra(mantra)

Deletes a Mantra.

Examples

iex> delete_mantra(mantra)
{:ok, %Mantra{}}

iex> delete_mantra(mantra)
{:error, %Ecto.Changeset{}}

delete_search_term(search_term)

Deletes a SearchTerm.

Examples

iex> delete_search_term(search_term)
{:ok, %SearchTerm{}}

iex> delete_search_term(search_term)
{:error, %Ecto.Changeset{}}

delete_tabs(tabs)

Deletes a Tabs.

Examples

iex> delete_tabs(tabs)
{:ok, %Tabs{}}

iex> delete_tabs(tabs)
{:error, %Ecto.Changeset{}}

delete_todo(todo)

Deletes a Todo.

Examples

iex> delete_todo(todo)
{:ok, %Todo{}}

iex> delete_todo(todo)
{:error, %Ecto.Changeset{}}

delete_user(user)

Deletes a User.

Examples

iex> delete_user(user)
{:ok, %User{}}

iex> delete_user(user)
{:error, %Ecto.Changeset{}}

get_attention!(id)

Gets a single mantra.

Raises Ecto.NoResultsError if the Mantra does not exist.

Examples

iex> get_mantra!(123)
%Mantra{}

iex> get_mantra!(456)
** (Ecto.NoResultsError)

get_category!(id)

Gets a single category.

Raises Ecto.NoResultsError if the Category does not exist.

Examples

iex> get_category!(123)
%Category{}

iex> get_category!(456)
** (Ecto.NoResultsError)

get_credential!(id)

Gets a single credential.

Raises Ecto.NoResultsError if the Credential does not exist.

Examples

iex> get_credential!(123)
%Credential{}

iex> get_credential!(456)
** (Ecto.NoResultsError)

get_mantra!(id)

Gets a single mantra.

Raises Ecto.NoResultsError if the Mantra does not exist.

Examples

iex> get_mantra!(123)
%Mantra{}

iex> get_mantra!(456)
** (Ecto.NoResultsError)

get_search_term!(id)

Gets a single search_term.

Raises Ecto.NoResultsError if the SearchTerm does not exist.

Examples

iex> get_search_term!(123)
%SearchTerm{}

iex> get_search_term!(456)
** (Ecto.NoResultsError)

get_tabs!(id)

Gets a single tabs.

Raises Ecto.NoResultsError if the Tabs does not exist.

Examples

iex> get_tabs!(123)
%Tabs{}

iex> get_tabs!(456)
** (Ecto.NoResultsError)

get_todo!(id)

Gets a single todo.

Raises Ecto.NoResultsError if the Todo does not exist.

Examples

iex> get_todo!(123)
%Todo{}

iex> get_todo!(456)
** (Ecto.NoResultsError)

get_user!(id)

Gets a single user.

Raises Ecto.NoResultsError if the User does not exist.

Examples

iex> get_user!(123)
%User{}

iex> get_user!(456)
** (Ecto.NoResultsError)

get_user_from_credential_info(map)

get_user_from_username(username)

list_categories()

Returns the list of categories.

Examples

iex> list_categories()
[%Category{}, ...]

list_credentials()

Returns the list of credentials.

Examples

iex> list_credentials()
[%Credential{}, ...]

list_mantras()

Returns the list of mantras.

Examples

iex> list_mantras()
[%Mantra{}, ...]

list_search_terms()

Returns the list of SearchTerms.

Examples

iex> list_search_term()
[%SearchTerm{}, ...]

list_tabs()

Returns the list of tabs.

Examples

iex> list_tabs()
[%Tabs{}, ...]

list_todos()

Returns the list of todos.

Examples

iex> list_todos()
[%Todo{}, ...]

list_user_attentions(user_id)

⚠️ this is not correct comment

Returns the list of mantras for a particular user.

Examples

iex> list_user_todos(%User{})
[%{id: integer, name: "my trace"}, ...]

list_user_mantras(user_id)

Returns the list of mantras for a particular user.

Examples

iex> list_user_todos(%User{})
[%{id: integer, name: "my trace"}, ...]

list_user_search_terms(user_id)

list_user_tabs(user_id)

list_user_todos(user_id)

Returns the list of todos for a particular user.

Examples

iex> list_user_todos(%User{})
[%{id: integer, name: "my trace"}, ...]

list_users()

Returns the list of users.

Examples

iex> list_users()
[%User{}, ...]

update_attention(attention, attrs)

Updates a mantra.

Examples

iex> update_mantra(mantra, %{field: new_value})
{:ok, %Mantra{}}

iex> update_mantra(mantra, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_category(category, attrs)

Updates a category.

Examples

iex> update_category(category, %{field: new_value})
{:ok, %Category{}}

iex> update_category(category, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_credential(credential, attrs)

Updates a credential.

Examples

iex> update_credential(credential, %{field: new_value})
{:ok, %Credential{}}

iex> update_credential(credential, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_mantra(mantra, attrs)

Updates a mantra.

Examples

iex> update_mantra(mantra, %{field: new_value})
{:ok, %Mantra{}}

iex> update_mantra(mantra, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_search_term(search_term, attrs)

Updates a search_term.

Examples

iex> update_search_term(search_term, %{field: new_value})
{:ok, %SearchTerm{}}

iex> update_search_term(search_term, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_tabs(tabs, attrs)

Updates a tabs.

Examples

iex> update_tabs(tabs, %{field: new_value})
{:ok, %Tabs{}}

iex> update_tabs(tabs, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_todo(todo, attrs)

Updates a todo.

Examples

iex> update_todo(todo, %{field: new_value})
{:ok, %Todo{}}

iex> update_todo(todo, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_user(user, attrs)

Updates a user.

Examples

iex> update_user(user, %{field: new_value})
{:ok, %User{}}

iex> update_user(user, %{field: bad_value})
{:error, %Ecto.Changeset{}}