flambe v0.0.1 Flambe.Traces

The Traces context.

Functions

change_event(event)

Returns an %Ecto.Changeset{} for tracking event changes.

Examples

iex> change_event(event)
%Ecto.Changeset{source: %Event{}}

change_thread(thread)

Returns an %Ecto.Changeset{} for tracking thread changes.

Examples

iex> change_thread(thread)
%Ecto.Changeset{source: %Thread{}}

change_trace(trace)

Returns an %Ecto.Changeset{} for tracking trace changes.

Examples

iex> change_trace(trace)
%Ecto.Changeset{source: %Trace{}}

create_activity(thread_id, attrs \\ %{})

Creates a activity.

Examples

iex> create_activity(%{field: value})
{:ok, %Activity{}}

iex> create_activity(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_event(trace_id, attrs \\ %{})

Creates a event.

Examples

iex> create_event(%{field: value})
{:ok, %Event{}}

iex> create_event(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_event(trace_id, activity_id, attrs)

Creates an event that is tied to a particular activity.

Examples

iex> create_event(%{field: value})
{:ok, %Event{}}

iex> create_event(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_thread(trace_id, attrs \\ %{})

Creates a thread.

Examples

iex> create_thread(%{field: value})
{:ok, %Thread{}}

iex> create_thread(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_trace(user, attrs \\ %{})

Creates a trace.

Examples

iex> create_trace(%{field: value})
{:ok, %Trace{}}

iex> create_trace(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_activity(activity)

Deletes a Activity.

Examples

iex> delete_activity(activity)
{:ok, %Activity{}}

iex> delete_activity(activity)
{:error, %Ecto.Changeset{}}

delete_event(event)

Deletes a Event.

Examples

iex> delete_event(event)
{:ok, %Event{}}

iex> delete_event(event)
{:error, %Ecto.Changeset{}}

delete_thread(thread)

Deletes a Thread.

Examples

iex> delete_thread(thread)
{:ok, %Thread{}}

iex> delete_thread(thread)
{:error, %Ecto.Changeset{}}

delete_trace(trace)

Deletes a Trace.

Examples

iex> delete_trace(trace)
{:ok, %Trace{}}

iex> delete_trace(trace)
{:error, %Ecto.Changeset{}}

get_activity!(id)

Gets a single activity.

Raises Ecto.NoResultsError if the Activity does not exist.

Examples

iex> get_activity!(123)
%Activity{}

iex> get_activity!(456)
** (Ecto.NoResultsError)

get_event!(id)

Gets a single event.

Raises Ecto.NoResultsError if the Event does not exist.

Examples

iex> get_event!(123)
%Event{}

iex> get_event!(456)
** (Ecto.NoResultsError)

get_thread!(id)

Gets a single thread.

Raises Ecto.NoResultsError if the Thread does not exist.

Examples

iex> get_thread!(123)
%Thread{}

iex> get_thread!(456)
** (Ecto.NoResultsError)

get_trace!(id)

Gets a single trace.

Raises Ecto.NoResultsError if the Trace does not exist.

Examples

iex> get_trace!(123)
%Trace{}

iex> get_trace!(456)
** (Ecto.NoResultsError)

get_trace_with_events(id)

list_activities()

Returns the list of activities.

Examples

iex> list_activities()
[%Activity{}, ...]

list_events()

Returns the list of events.

Examples

iex> list_events()
[%Event{}, ...]

list_threads()

Returns the list of threads.

Examples

iex> list_threads()
[%Thread{}, ...]

list_trace_events(trace)

Returns the list of events for a particular trace.

Examples

iex> list_trace_events(%Trace{})
[%Event{}, ...]

list_trace_threads(trace)

list_traces()

Returns the list of traces.

Examples

iex> list_traces()
[%Trace{}, ...]

list_user_traces(user_id)

Returns the list of traces for a particular user.

Examples

iex> list_user_traces(%User{})
[%{id: integer, name: "my trace"}, ...]

update_activity(activity, attrs)

Updates a activity.

Examples

iex> update_activity(activity, %{field: new_value})
{:ok, %Activity{}}

iex> update_activity(activity, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_event(event, attrs)

Updates a event.

Examples

iex> update_event(event, %{field: new_value})
{:ok, %Event{}}

iex> update_event(event, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_thread(thread, attrs)

Updates a thread.

Examples

iex> update_thread(thread, %{field: new_value})
{:ok, %Thread{}}

iex> update_thread(thread, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_trace(trace, attrs)

Updates a trace.

Examples

iex> update_trace(trace, %{field: new_value})
{:ok, %Trace{}}

iex> update_trace(trace, %{field: bad_value})
{:error, %Ecto.Changeset{}}