flambe v0.0.1 SteadyWeb.Router.Helpers

Module with named helpers generated from SteadyWeb.Router.

Functions

activity_path(conn_or_endpoint, action)

activity_path(conn_or_endpoint, action, id)

activity_path(conn_or_endpoint, action, id, params)

activity_url(conn_or_endpoint, action)

activity_url(conn_or_endpoint, action, id)

activity_url(conn_or_endpoint, action, id, params)

attention_path(conn_or_endpoint, action)

attention_path(conn_or_endpoint, action, params)

attention_path(conn_or_endpoint, action, id, params)

attention_url(conn_or_endpoint, action)

attention_url(conn_or_endpoint, action, params)

attention_url(conn_or_endpoint, action, id, params)

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, provider)

auth_path(conn_or_endpoint, action, provider, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, provider)

auth_url(conn_or_endpoint, action, provider, params)

category_path(conn_or_endpoint, action)

category_path(conn_or_endpoint, action, params)

category_path(conn_or_endpoint, action, id, params)

category_url(conn_or_endpoint, action)

category_url(conn_or_endpoint, action, params)

category_url(conn_or_endpoint, action, id, params)

cms_page_path(conn_or_endpoint, action)

cms_page_path(conn_or_endpoint, action, params)

cms_page_path(conn_or_endpoint, action, id, params)

cms_page_url(conn_or_endpoint, action)

cms_page_url(conn_or_endpoint, action, params)

cms_page_url(conn_or_endpoint, action, id, params)

event_path(conn_or_endpoint, action)

event_path(conn_or_endpoint, action, params)

event_path(conn_or_endpoint, action, id, params)

event_url(conn_or_endpoint, action)

event_url(conn_or_endpoint, action, params)

event_url(conn_or_endpoint, action, id, params)

mantra_path(conn_or_endpoint, action)

mantra_path(conn_or_endpoint, action, params)

mantra_path(conn_or_endpoint, action, id, params)

mantra_url(conn_or_endpoint, action)

mantra_url(conn_or_endpoint, action, params)

mantra_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

search_term_path(conn_or_endpoint, action)

search_term_path(conn_or_endpoint, action, params)

search_term_path(conn_or_endpoint, action, id, params)

search_term_url(conn_or_endpoint, action)

search_term_url(conn_or_endpoint, action, params)

search_term_url(conn_or_endpoint, action, id, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

tabs_path(conn_or_endpoint, action)

tabs_path(conn_or_endpoint, action, params)

tabs_path(conn_or_endpoint, action, id, params)

tabs_url(conn_or_endpoint, action)

tabs_url(conn_or_endpoint, action, params)

tabs_url(conn_or_endpoint, action, id, params)

thread_path(conn_or_endpoint, action)

thread_path(conn_or_endpoint, action, params)

thread_path(conn_or_endpoint, action, id, params)

thread_url(conn_or_endpoint, action)

thread_url(conn_or_endpoint, action, params)

thread_url(conn_or_endpoint, action, id, params)

todo_path(conn_or_endpoint, action)

todo_path(conn_or_endpoint, action, params)

todo_path(conn_or_endpoint, action, id, params)

todo_url(conn_or_endpoint, action)

todo_url(conn_or_endpoint, action, params)

todo_url(conn_or_endpoint, action, id, params)

trace_path(conn_or_endpoint, action)

trace_path(conn_or_endpoint, action, params)

trace_path(conn_or_endpoint, action, id, params)

trace_url(conn_or_endpoint, action)

trace_url(conn_or_endpoint, action, params)

trace_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)